Applied slugs to api paths
