+++ pages/api/search/[term].js
