+++ pages/api/users/index.js
@@ -1,40 +1,7 @@
-import {
-  getUser,
-  getAllUsers,
-  getApprovedUsers,
-  getNotApprovedUsers,
-  setUserApproval,
-  deleteUser,
-  saveUser,
-} from '../../../server/db';
+import { setUserApproval, deleteUser } from '../../../server/db';
 
 const getCollectionResults = async (req, res) => {
   let result;
-  const { user } = req.query;
-
-  if (user) {
-    switch (user) {
-      case 'approved':
-        result = await getApprovedUsers();
-        break;
-      case 'notapproved':
-        result = await getNotApprovedUsers();
-        break;
-      case 'getall':
-        result = await getAllUsers();
-        break;
-      default:
-        result = await getUser(user);
-        if (!result) {
-          const newUserDocument = {
-            email: user,
-            role: 100,
-            approved: false,
-          };
-          result = await saveUser(newUserDocument);
-        }
-    }
-  }
 
   if (req.method === 'POST') {
     const { user, approval, deleteuser } = req.body;
